Fix reloading issue
Fix issue on reloading (e.g. calling #reload! in rails console). It couldn't be done, due to the fact that reloaded module CoreExtensions::ActionDispatch::Routing had been included in ActionDispatch::Routing::Mapper, that was not reloaded (is part of an external gem).
This solution prevents ./lib directory from being reloaded and requires relevant files in initalizer for core_extensions.